refactor: extract InputDropdownSelect from NewBid
