+++ src/components/inputDropdownSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './InputDropdownSelect';
