--- src/routes/bids/newBid/NewBidForm.scss
@@ -1,62 +1,0 @@
-@import 'assets/scss/variables';
-
-.newBid {
-  .security__display {
-    font-family: 'Nexa Bold';
-    font-size: 1.2rem;
-    line-height: 1.2rem;
-    &--none {
-      font-family: 'Nexa Light';
-      opacity: 0.6;
-    }
-  }
-
-  .securities__dropdown {
-    height: 2.75rem;
-    padding-left: 0.75em !important;
-    border-top-left-radius: 10px !important;
-    border-bottom-left-radius: 10px !important;
-    min-width: 6rem;
-    background: $grey;
-
-    &.is-danger {
-      &:hover {
-        border-color: $red;
-      }
-      border-color: $red;
-
-      &:focus-within,
-      &:focus,
-      &:active {
-        border-color: $red;
-        box-shadow: 0 0 0 0.125em rgba(255, 56, 96, 0.25);
-      }
-    }
-
-    .react-dropdown-select {
-      &-dropdown {
-        width: max-content;
-        min-width: inherit;
-      }
-
-      &-content {
-        align-items: center;
-        height: inherit;
-        white-space: nowrap;
-
-        input {
-          text-align: start;
-        }
-
-        span {
-          height: inherit;
-          display: flex;
-          align-items: center;
-        }
-      }
-      &-item {
-        width: 100%;
-      }
-    }
-  }
-}
